+++ buildSrc/src/main/kotlin/Versions.kt
@@ -1,4 +1,4 @@
 object Versions {
     const val group = "com.github.useradgents.skot"
-    const val version = "1.3.6-ua"
+    const val version = "1.3.7-ua"
 }