+++ plugin/src/main/kotlin/tech/skot/Versions.kt
@@ -1,19 +1,25 @@
 package tech.skot
 
+import kotlin.Int
+import kotlin.String
+
+/**
+ * This code is generated from build.gradle.kts
+ */
 public object Versions {
-  public const val skot: String = "1.2.41-ua"
+  public const val skot: String = "1.3.7-ua"
 
   public const val group: String = "com.github.useradgents.skot"
 
-  public const val serialization: String = "1.7.3"
+  public const val serialization: String = "1.8.0"
 
   public const val kotlinxDateTime: String = "0.6.1"
 
-  public const val ktor: String = "2.3.12"
+  public const val ktor: String = "3.0.3"
 
-  public const val kotlin: String = "2.0.20"
+  public const val kotlin: String = "2.1.10"
 
-  public const val kotlinCoroutines: String = "1.9.0"
+  public const val kotlinCoroutines: String = "1.10.1"
 
   public const val kotlinpoet: String = "1.18.1"
 
@@ -27,7 +33,7 @@ public object Versions {
 
   public const val sqldelight: String = "1.5.5"
 
-  public const val desugaring: String = "2.1.2"
+  public const val desugaring: String = "2.1.4"
 
   public const val leakcanary: String = "2.14"